requirements.txt
numpy<2
sounddevice
scipy
openai-whisper
deep-translator
certifi
tk